+++ genmon.py
@@ -36,7 +36,7 @@ except Exception as e1:
     print("Error: " + str(e1))
     sys.exit(2)
 
-GENMON_VERSION = "V1.18.00"
+GENMON_VERSION = "V1.18.01"
 
 #------------ Monitor class ----------------------------------------------------
 class Monitor(MySupport):