+++ genserv.py
@@ -625,6 +625,31 @@ def GetAddOns():
             "Number to send SMS message from. This should be a twilio phone number.",
             bounds = 'required InternationalPhone',
             display_name = "Twilio Phone Number")
+        AddOnCfg['gensms']['parameters']['notify_error'] = CreateAddOnParam(
+            ConfigFiles[GENSMS_CONFIG].ReadValue("notify_error", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for errors (Generator Alarms).",
+            display_name = "Notify for Errors")
+        AddOnCfg['gensms']['parameters']['notify_warn'] = CreateAddOnParam(
+            ConfigFiles[GENSMS_CONFIG].ReadValue("notify_warn", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for warnings (service due, fuel low).",
+            display_name = "Notify for Warnings")
+        AddOnCfg['gensms']['parameters']['notify_info'] = CreateAddOnParam(
+            ConfigFiles[GENSMS_CONFIG].ReadValue("notify_info", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for information (switch state change, engine state change, exercising).",
+            display_name = "Notify for Information")
+        AddOnCfg['gensms']['parameters']['notify_outage'] = CreateAddOnParam(
+            ConfigFiles[GENSMS_CONFIG].ReadValue("notify_outage", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for outages.",
+            display_name = "Notify for Outages")
+        AddOnCfg['gensms']['parameters']['notify_sw_update'] = CreateAddOnParam(
+            ConfigFiles[GENSMS_CONFIG].ReadValue("notify_sw_update", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for software updates.",
+            display_name = "Notify for Software Updates")
 
         #GENSMS_MODEM
         AddOnCfg['gensms_modem'] = collections.OrderedDict()
@@ -659,6 +684,33 @@ def GetAddOns():
             'boolean',
             "Enable to log at commands to the log file.",
             display_name = "Log AT Commands")
+
+        AddOnCfg['gensms_modem']['parameters']['notify_error'] = CreateAddOnParam(
+            ConfigFiles[MYMODEM_CONFIG].ReadValue("notify_error", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for errors (Generator Alarms).",
+            display_name = "Notify for Errors")
+        AddOnCfg['gensms_modem']['parameters']['notify_warn'] = CreateAddOnParam(
+            ConfigFiles[MYMODEM_CONFIG].ReadValue("notify_warn", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for warnings (service due, fuel low).",
+            display_name = "Notify for Warnings")
+        AddOnCfg['gensms_modem']['parameters']['notify_info'] = CreateAddOnParam(
+            ConfigFiles[MYMODEM_CONFIG].ReadValue("notify_info", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for information (switch state change, engine state change, exercising).",
+            display_name = "Notify for Information")
+        AddOnCfg['gensms_modem']['parameters']['notify_outage'] = CreateAddOnParam(
+            ConfigFiles[MYMODEM_CONFIG].ReadValue("notify_outage", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for outages.",
+            display_name = "Notify for Outages")
+        AddOnCfg['gensms_modem']['parameters']['notify_sw_update'] = CreateAddOnParam(
+            ConfigFiles[MYMODEM_CONFIG].ReadValue("notify_sw_update", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for software updates.",
+            display_name = "Notify for Software Updates")
+
         # modem type - select the type of modem used. For future use. Presently "LTEPiHat" is the only option
         #modem_type = LTEPiHat
 
@@ -689,6 +741,31 @@ def GetAddOns():
             "Notification sound identifier. See https://pushover.net/api#sounds for a full list of sound IDs",
             bounds = 'minmax:3:20',
             display_name = "Push Sound")
+        AddOnCfg['genpushover']['parameters']['notify_error'] = CreateAddOnParam(
+            ConfigFiles[GENPUSHOVER_CONFIG].ReadValue("notify_error", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for errors (Generator Alarms).",
+            display_name = "Notify for Errors")
+        AddOnCfg['genpushover']['parameters']['notify_warn'] = CreateAddOnParam(
+            ConfigFiles[GENPUSHOVER_CONFIG].ReadValue("notify_warn", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for warnings (service due, fuel low).",
+            display_name = "Notify for Warnings")
+        AddOnCfg['genpushover']['parameters']['notify_info'] = CreateAddOnParam(
+            ConfigFiles[GENPUSHOVER_CONFIG].ReadValue("notify_info", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for information (switch state change, engine state change, exercising).",
+            display_name = "Notify for Information")
+        AddOnCfg['genpushover']['parameters']['notify_outage'] = CreateAddOnParam(
+            ConfigFiles[GENPUSHOVER_CONFIG].ReadValue("notify_outage", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for outages.",
+            display_name = "Notify for Outages")
+        AddOnCfg['genpushover']['parameters']['notify_sw_update'] = CreateAddOnParam(
+            ConfigFiles[GENPUSHOVER_CONFIG].ReadValue("notify_sw_update", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for software updates.",
+            display_name = "Notify for Software Updates")
 
         # GENSYSLOG
         AddOnCfg['gensyslog'] = collections.OrderedDict()
@@ -832,6 +909,31 @@ def GetAddOns():
             "Use this to make the title of the message a link i.e. link to the genmon web interface.",
             bounds = 'HTTPAddress',
             display_name = "Title Link")
+        AddOnCfg['genslack']['parameters']['notify_error'] = CreateAddOnParam(
+            ConfigFiles[GENSLACK_CONFIG].ReadValue("notify_error", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for errors (Generator Alarms).",
+            display_name = "Notify for Errors")
+        AddOnCfg['genslack']['parameters']['notify_warn'] = CreateAddOnParam(
+            ConfigFiles[GENSLACK_CONFIG].ReadValue("notify_warn", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for warnings (service due, fuel low).",
+            display_name = "Notify for Warnings")
+        AddOnCfg['genslack']['parameters']['notify_info'] = CreateAddOnParam(
+            ConfigFiles[GENSLACK_CONFIG].ReadValue("notify_info", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for information (switch state change, engine state change, exercising).",
+            display_name = "Notify for Information")
+        AddOnCfg['genslack']['parameters']['notify_outage'] = CreateAddOnParam(
+            ConfigFiles[GENSLACK_CONFIG].ReadValue("notify_outage", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for outages.",
+            display_name = "Notify for Outages")
+        AddOnCfg['genslack']['parameters']['notify_sw_update'] = CreateAddOnParam(
+            ConfigFiles[GENSLACK_CONFIG].ReadValue("notify_sw_update", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for software updates.",
+            display_name = "Notify for Software Updates")
 
         # GENEXERCISE
         ControllerInfo = GetControllerInfo("controller").lower()
@@ -913,6 +1015,31 @@ def GetAddOns():
             "Email to SMS email recipient. Must be a valid email address",
             bounds = 'required email',
             display_name = "Email to SMS address")
+        AddOnCfg['genemail2sms']['parameters']['notify_error'] = CreateAddOnParam(
+            ConfigFiles[GENEMAIL2SMS_CONFIG].ReadValue("notify_error", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for errors (Generator Alarms).",
+            display_name = "Notify for Errors")
+        AddOnCfg['genemail2sms']['parameters']['notify_warn'] = CreateAddOnParam(
+            ConfigFiles[GENEMAIL2SMS_CONFIG].ReadValue("notify_warn", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for warnings (service due, fuel low).",
+            display_name = "Notify for Warnings")
+        AddOnCfg['genemail2sms']['parameters']['notify_info'] = CreateAddOnParam(
+            ConfigFiles[GENEMAIL2SMS_CONFIG].ReadValue("notify_info", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for information (switch state change, engine state change, exercising).",
+            display_name = "Notify for Information")
+        AddOnCfg['genemail2sms']['parameters']['notify_outage'] = CreateAddOnParam(
+            ConfigFiles[GENEMAIL2SMS_CONFIG].ReadValue("notify_outage", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for outages.",
+            display_name = "Notify for Outages")
+        AddOnCfg['genemail2sms']['parameters']['notify_sw_update'] = CreateAddOnParam(
+            ConfigFiles[GENEMAIL2SMS_CONFIG].ReadValue("notify_sw_update", return_type = bool, default = True),
+            'boolean',
+            "Send mssages for software updates.",
+            display_name = "Notify for Software Updates")
 
         #GENTANKUTIL
         AddOnCfg['gentankutil'] = collections.OrderedDict()
@@ -1373,6 +1500,10 @@ def ReadAdvancedSettingsFromFile():
         ConfigSettings["min_outage_duration"] = ['int', 'Minimum Outage Duration', 64, "0", "", 0, GENMON_CONFIG, GENMON_SECTION,"min_outage_duration"]
         ConfigSettings["multi_instance"] = ['boolean', 'Allow Multiple Genmon Instances', 65, False, "", 0, GENMON_CONFIG, GENMON_SECTION, "multi_instance"]
 
+        if GStartInfo["Linux"]:
+            ConfigSettings["uselinuxwifisignalgauge"] = ['boolean', 'Show Wifi Signal Strength Gauge', 60, True, "", 0, GENMON_CONFIG, GENMON_SECTION, "uselinuxwifisignalgauge"]
+        if GStartInfo["RaspbeerryPi"]:
+            ConfigSettings["useraspberrypicputempgauge"] = ['boolean', 'Show CPU Temperature Gauge', 60, True, "", 0, GENMON_CONFIG, GENMON_SECTION, "useraspberrypicputempgauge"] 
 
         for entry, List in ConfigSettings.items():
             if List[6] == GENMON_CONFIG: